NCW cron node checks

+++ web/modules/custom/ncw_migration/src/NmCron.php
@@ -100,7 +100,9 @@ class NmCron implements ContainerInjectionInterface{
         $migration->getIdMap()->deleteDestination(['nid' => $del['destid1']]);
         // Remove the node.
         $node = $this->entityTypeManager->getStorage('node')->load($del['destid1']);
-        $node->delete();
+        if (isset($node)) {
+          $node->delete();
+        }
       }
     }
   }
